refactored iptv logic
