+++ src/modules/iptv/parsers/index.ts
@@ -1,0 +1,2 @@
+export { fromPlaylistLine } from './playlist-parser';
+export { parseXMLTV } from './xmltv-parser';
